requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
gspread==6.1.2
google-generativeai==0.8.3
python-multipart==0.0.9
